Time: 1 ms (99.61%) | Memory: 56.8 MB (89.03%) - LeetSync

+++ 53-maximum-subarray/maximum-subarray.java
@@ -1,7 +1,7 @@
 class Solution {
     public int maxSubArray(int[] nums) {
         int n = nums.length;
-        long maximum = Long.MIN_VALUE;
+        int maximum = Integer.MIN_VALUE;
         int sum = 0;
       for (int i = 0; i < n; i++) {
